I did not realize that I needed that

+++ labs/lab06/src/main/scala/example/App.scala
@@ -92,7 +92,9 @@ object App {
         // ._4 is the total for that record
         // ._5 is the state
         (id, records.head._5, 
-          records.foldleft(0.0)({ case (total, (_, _, _, t, _)) => total + t}))
+          records.foldleft(0.0)({ 
+            case (total, (_, _, _, t, _)) => total + t.toDouble
+          }))
     })
 
     // =========================================================================
